refactor: split node-only browser launchers out of the fallback generators
`generatePlayerHtml` / `generateStrudelPlayerHtml` are pure string builders, but
they shared a module with `openPlayerInBrowser` / `openStrudelInBrowser`, whose
node:child_process / node:crypto / node:fs / node:os / node:path imports made the
generators unreachable from the Cloudflare Worker without dragging unenv's Node
polyfill set into its bundle (+1.1 MB raw, +213 KiB gzip of code that can never
run there — measured with `wrangler deploy --dry-run`).

The launchers move to src/open-in-browser.ts (the two near-identical
write-file-then-shell-out bodies collapse into one `writeAndOpen` helper); the
two fallback modules keep their generators and are now runtime-agnostic, which
is what the upcoming Worker /play + /score routes need. server.ts is the only
consumer of the launchers and just re-points two imports.
