+++ src/browser-fallback.ts
@@ -2,11 +2,9 @@
  * Browser fallback for non-UI MCP clients.
  * Generates a self-contained HTML player and opens it in the default browser.
  */
-import { execFile } from "node:child_process";
-import { randomUUID } from "node:crypto";
-import fs from "node:fs/promises";
-import path from "node:path";
-import os from "node:os";
+// Runtime-agnostic: no node: imports here, so the Cloudflare Worker can render
+// the same page for its /score route. The disk-writing + browser-launching half
+// lives in src/open-in-browser.ts.
 import {
   INSTRUMENTS,
   STYLE_PRESETS,
@@ -436,31 +434,4 @@ export function generatePlayerHtml(options: BrowserPlayerOptions): string {
   </script>
 </body>
 </html>`;
-}
-
-export async function openPlayerInBrowser(
-  options: BrowserPlayerOptions,
-  outputDir?: string,
-): Promise<string> {
-  const html = generatePlayerHtml(options);
-
-  const outDir = outputDir ?? path.join(os.homedir(), "Desktop", "mcp-music-studio");
-  await fs.mkdir(outDir, { recursive: true });
-
-  const filename = `player-${randomUUID()}.html`;
-  const filepath = path.join(outDir, filename);
-  await fs.writeFile(filepath, html, "utf-8");
-
-  // Try to open in default browser (may fail in sandboxed environments).
-  // Use execFile with an argv array so no shell parses the path.
-  const platform = process.platform;
-  const openErr = (err: Error | null) => {
-    if (err) console.error("Failed to open browser:", err.message);
-  };
-  if (platform === "darwin") execFile("open", [filepath], openErr);
-  else if (platform === "win32")
-    execFile("cmd", ["/c", "start", "", filepath], openErr);
-  else execFile("xdg-open", [filepath], openErr);
-
-  return filepath;
 }
